# sync-server/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Fractonica API
  version: current
  summary: Server-authoritative journal, event, and synchronization API.
  description: |
    The Fractonica API is the canonical server contract.
    It is multi-user, cursor-synchronized, and intended for the first-party iOS
    client as well as user-authorized automation agents.

    Authentication uses either a short-lived JWT access token (`JwtBearer`) or
    a user-issued API key (`ApiKeyBearer`) in the `Authorization: Bearer ...`
    header. API keys are bound to one user and one logical device. Operations
    carrying `x-required-scopes` enforce those scopes for API-key callers; JWT
    sessions act with the authenticated user's full authority.

    Production deployments must expose the API only over HTTPS/TLS and reject
    plain HTTP at the edge. Passwords and bearer credentials are otherwise
    observable in transit. Plain HTTP is for loopback development only.

    Records are public by default. Public record payloads are readable and may
    be rendered from a template. Private record payloads and private
    media are encrypted by the client before upload; the server stores only an
    authenticated ciphertext envelope and intentionally cannot render, index,
    or embed their plaintext.

    All timestamps are RFC 3339 instants. Record API identifiers are globally
    unique five-character Base64URL strings; their owner-only `originId` UUIDs
...
    are UUIDs.
...

...
    immediately retrying.

...
  license:
...
servers:
...
security:
  - JwtBearer: []
  - ApiKeyBearer: []
...
paths:
...
  /sync/stats:
    get:
      operationId: getSyncStats
      tags: [Synchronization]
      summary: Count the authenticated user's restorable resources
      description: |
        Returns a lightweight owner-only snapshot for restore planning and
        progress indicators. `media.restorable` counts distinct ready public
        media objects attached to active public records; it is the media total
        a plaintext client can recover through the record collection. Capture
        `cursor` before paging those collections, then resume the change feed
        from it so writes concurrent with the snapshot are not missed.
      x-required-scopes: [sync:read]
      responses:
        '200':
          description: Current active resource and media-byte totals.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncStats'
              example:
                cursor: eyJ2IjoxLCJraW5kIjoic3luYy1jaGFuZ2VzIiwic2VxIjoiNDgxMiJ9
                records:
                  total: 728
                  public: 728
                  private: 0
                  pastTera: 48
                  pastGiga: 7
                  pastMega: 1
                events: { total: 0 }
                tags: { total: 14 }
                templates: { total: 2 }
                media:
                  total: 1936
                  byteLength: 4096432128
                  photo: 1712
                  video: 62
                  audio: 162
                  restorable: 1936
                  restorableByteLength: 4096432128
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /sync/changes:
    get:
      operationId: getSyncChanges
      tags: [Synchronization]
      summary: Read the authenticated user's ordered change feed
      description: |
        Before the first request, fully reconcile the authoritative owner
        collections the client stores; a request without `cursor` is a feed
        anchor, not a replacement for that snapshot after change retention has
        pruned older rows. Then omit `cursor`, apply changes in returned order,
        and persist `nextCursor` only after the entire page commits locally.
        Upserts contain the complete current resource. Deletes contain
        tombstones. Cursors are user- and query-bound; using one with another
        user or a different resource filter returns `400`.
      x-required-scopes: [sync:read]
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: resourceType
          in: query
          description: Repeat to include selected resource types; omit for all.
          style: form
          explode: true
          schema:
            type: array
            maxItems: 8
            uniqueItems: true
            items:
              $ref: '#/components/schemas/SyncResourceType'
      responses:
        '200':
          description: Ordered changes and the cursor after the last returned sequence.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangePage'
              example:
                data:
                  - sequence: 4812
                    changedAt: '2026-07-14T18:04:12Z'
                    resourceType: event
                    operation: upsert
                    resourceId: 20a78723-c33e-4794-a036-5da69a15e8bf
                    revision: 1
                    etag: '"event-20a78723-c33e-4794-a036-5da69a15e8bf-r1"'
                    resource:
                      id: 20a78723-c33e-4794-a036-5da69a15e8bf
                      userId: e42b4fde-8baf-4b95-8bc8-5395b68d0dd2
                      author:
                        id: e42b4fde-8baf-4b95-8bc8-5395b68d0dd2
                        login: sun
                        displayName: Sun
                        sarosAnchor: 141
                      deviceId: 2dca8eab-00a8-4e94-9bd2-2fcbfe17e890
                      visibility: public
                      startsAt: '2026-07-14T16:42:00Z'
                      endsAt: '2026-07-14T16:55:00Z'
                      label: X1.7 solar flare
                      type: 1001
                      metadata: {source: noaa-swpc}
                      references: []
                      revision: 1
                      createdAt: '2026-07-14T18:04:12Z'
                      updatedAt: '2026-07-14T18:04:12Z'
                nextCursor: eyJzZXEiOjQ4MTJ9
                hasMore: false
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '410':
          $ref: '#/components/responses/CursorExpired'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /sync/batches:
    post:
      operationId: applySyncBatch
      tags: [Synchronization]
      summary: Apply a bounded batch of idempotent client mutations
      description: |
        Mutations are evaluated in array order. Each carries a stable
        `clientMutationId`. With `atomic: false` (the default), failures are
        reported per item and do not roll back successful siblings. With
        `atomic: true`, any failure rolls back the full batch and returns `409`.
        Successful writes enter the same change feed as direct API writes.
      x-required-scopes: [sync:write]
      x-body-limit-bytes: 16777216
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncBatchRequest'
      responses:
        '200':
          description: Per-mutation outcomes for a non-atomic batch.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncBatchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '422':
          $ref: '#/components/responses/UnprocessableContent'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    JwtBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Short-lived Ed25519-signed access JWT issued by login or refresh. The
        token includes `iss`, `aud`, `sub`, `exp`, and `kid` claims.
    ApiKeyBearer:
      type: http
      scheme: bearer
      bearerFormat: Fractonica API key
      description: |
        User-issued secret beginning with `exk_`. Keys are stored hashed,
        device-bound, scoped, independently expirable, and revocable. The
        plaintext secret appears only in the initial creation response; a replay
        returns the created `apiKeyId`, never the secret.
  parameters:
...
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor returned by the immediately preceding page.
      schema:
        type: string
        minLength: 1
        maxLength: 2048
    Limit:
      name: limit
      in: query
      description: Maximum number of items to return.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Caller-generated identifier for one logical mutation. It must be reused
        for retries of the same body and must not be reused for another body.
        Most operations replay their original response. API-key creation is the
        deliberate exception: replay returns a 409 problem containing the
        created `apiKeyId` because its plaintext secret is returned only once.
      schema:
        type: string
        minLength: 8
        maxLength: 255
        pattern: '^[\x21-\x7E]+$'
      example: 01J2V7YCB9Q3WZ2P6AQ03M6DN8
...
  headers:
...
    RequestId:
      description: Server-assigned request correlation identifier.
      schema:
        type: string
        minLength: 1
        maxLength: 128
  responses:
    BadRequest:
      description: The request syntax, parameter combination, or cursor is invalid.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://api.exeligmos.app/problems/invalid-request
            title: Invalid request
            status: 400
            detail: The occurredAfter timestamp must precede occurredBefore.
            code: invalid_request
            requestId: req_01J2V8E3VBA5QYFAE6BX27DE5Z
    Unauthorized:
      description: Authentication is missing, expired, revoked, or invalid.
      headers:
        WWW-Authenticate:
          schema:
            type: string
          example: Bearer realm="exeligmos"
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: The credential lacks the scope or ownership required for this operation.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    Conflict:
      description: The mutation conflicts with current state or idempotency history.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    CursorExpired:
      description: The cursor predates retained change history; perform a full reconciliation.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://api.exeligmos.app/problems/cursor-expired
            title: Synchronization cursor expired
            status: 410
            detail: Perform a full collection reconciliation before restarting the change feed.
            code: cursor_expired
            requestId: req_01J2V8E3VBA5QYFAE6BX27DE5Z
...
    PayloadTooLarge:
      description: The request or declared media object exceeds configured bounds.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    UnprocessableContent:
      description: The request is syntactically valid but violates domain constraints.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ValidationProblem'
    ServiceUnavailable:
      description: A database statement or lock exceeded its safety deadline; the operation may be retried.
      headers:
        Retry-After:
          description: Minimum number of seconds to wait before retrying.
          schema:
            type: integer
            minimum: 1
          example: 1
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: urn:exeligmos:problem:database-timeout
            title: Service Unavailable
            status: 503
            detail: The database could not complete the request within its safety deadline.
            code: database_timeout
            requestId: req_01J2V8E3VBA5QYFAE6BX27DE5Z
    TooManyRequests:
      description: Rate or concurrency limit exceeded.
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 1
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 Problem Details with stable application extensions.
      required: [type, title, status, code, requestId]
      properties:
        type:
          type: string
          format: uri-reference
          description: Stable problem type URI.
        title:
          type: string
          minLength: 1
          maxLength: 200
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
          maxLength: 4000
        instance:
          type: string
          format: uri-reference
        code:
          type: string
          pattern: '^[a-z][a-z0-9_]{1,63}$'
        requestId:
          type: string
          minLength: 1
          maxLength: 128
      additionalProperties: true
...
    ValidationProblem:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          required: [errors]
          properties:
            errors:
              type: array
              minItems: 1
              maxItems: 100
              items:
                $ref: '#/components/schemas/FieldError'
...
    CursorPageMeta:
      type: object
      required: [hasMore]
      properties:
        nextCursor:
          type: string
          minLength: 1
          maxLength: 2048
        hasMore:
          type: boolean
...
    ResourceIdentifier:
      description: >-
        Record resources use RecordPublicId; all other resource types use UUIDs.
      oneOf:
        - $ref: '#/components/schemas/RecordPublicId'
        - type: string
          format: uuid
...
    SyncResourceCount:
      type: object
      required: [total]
      properties:
        total:
          type: integer
          format: int64
          minimum: 0
      additionalProperties: false
    SyncStats:
      type: object
      required: [cursor, records, events, tags, templates, media]
      properties:
        cursor:
          type: string
          minLength: 2
          description: Opaque all-resource change cursor captured with these totals.
        records:
          type: object
          required: [total, public, private, pastTera, pastGiga, pastMega]
          properties:
            total: { type: integer, format: int64, minimum: 0 }
            public: { type: integer, format: int64, minimum: 0 }
            private: { type: integer, format: int64, minimum: 0 }
            pastTera: { type: integer, format: int64, minimum: 0 }
            pastGiga: { type: integer, format: int64, minimum: 0 }
            pastMega: { type: integer, format: int64, minimum: 0 }
          additionalProperties: false
        events:
          $ref: '#/components/schemas/SyncResourceCount'
        tags:
          $ref: '#/components/schemas/SyncResourceCount'
        templates:
          $ref: '#/components/schemas/SyncResourceCount'
        media:
          type: object
          required: [total, byteLength, photo, video, audio, restorable, restorableByteLength]
          properties:
            total: { type: integer, format: int64, minimum: 0 }
            byteLength: { type: integer, format: int64, minimum: 0 }
            photo: { type: integer, format: int64, minimum: 0 }
            video: { type: integer, format: int64, minimum: 0 }
            audio: { type: integer, format: int64, minimum: 0 }
            restorable: { type: integer, format: int64, minimum: 0 }
            restorableByteLength: { type: integer, format: int64, minimum: 0 }
          additionalProperties: false
      additionalProperties: false
    SyncResourceType:
      type: string
      enum: [record, event, tag, template, device, media, user, subscription]
    ChangeOperation:
      type: string
      enum: [upsert, delete]
    Tombstone:
      type: object
      required: [id, userId, resourceType, revision, deletedAt]
      properties:
        id:
          $ref: '#/components/schemas/ResourceIdentifier'
        userId:
          type: string
          format: uuid
        resourceType:
          $ref: '#/components/schemas/SyncResourceType'
        revision:
          type: integer
          format: int64
          minimum: 1
        deletedAt:
          type: string
          format: date-time
      additionalProperties: false
    SyncResource:
      oneOf:
        - $ref: '#/components/schemas/Record'
        - $ref: '#/components/schemas/Event'
        - $ref: '#/components/schemas/Tag'
        - $ref: '#/components/schemas/Template'
        - $ref: '#/components/schemas/Device'
        - $ref: '#/components/schemas/MediaObject'
        - $ref: '#/components/schemas/User'
        - $ref: '#/components/schemas/Subscription'
    Change:
      type: object
      required:
        - sequence
        - changedAt
        - resourceType
        - operation
        - resourceId
        - revision
        - etag
      properties:
        sequence:
          type: integer
          format: int64
          minimum: 1
        changedAt:
          type: string
          format: date-time
        resourceType:
          $ref: '#/components/schemas/SyncResourceType'
        operation:
          $ref: '#/components/schemas/ChangeOperation'
        resourceId:
          $ref: '#/components/schemas/ResourceIdentifier'
        revision:
          type: integer
          format: int64
          minimum: 1
        etag:
          type: string
          minLength: 3
        resource:
          $ref: '#/components/schemas/SyncResource'
        tombstone:
          $ref: '#/components/schemas/Tombstone'
      oneOf:
        - properties:
            operation:
              const: upsert
            resource:
              $ref: '#/components/schemas/SyncResource'
            tombstone:
              $ref: '#/components/schemas/Tombstone'
          required: [resource]
          not:
            properties:
              tombstone:
                $ref: '#/components/schemas/Tombstone'
            required: [tombstone]
        - properties:
            operation:
              const: delete
            resource:
              $ref: '#/components/schemas/SyncResource'
            tombstone:
              $ref: '#/components/schemas/Tombstone'
          required: [tombstone]
          not:
            properties:
              resource:
                $ref: '#/components/schemas/SyncResource'
            required: [resource]
      additionalProperties: false
    ChangePage:
      allOf:
        - $ref: '#/components/schemas/CursorPageMeta'
        - type: object
          required: [data, nextCursor]
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Change'
            nextCursor:
              type: string
              minLength: 1
              maxLength: 2048
...
    UpsertEventMutation:
      type: object
      required: [kind, clientMutationId, event]
      properties:
        kind:
          type: string
          const: upsertEvent
        clientMutationId:
          $ref: '#/components/schemas/MutationId'
        ifMatch:
          type: string
          description: Required when replacing an existing event.
        event:
          $ref: '#/components/schemas/CreateEventRequest'
      additionalProperties: false
...
    SyncMutation:
      oneOf:
        - $ref: '#/components/schemas/UpsertRecordMutation'
        - $ref: '#/components/schemas/UpsertEventMutation'
        - $ref: '#/components/schemas/UpsertTagMutation'
        - $ref: '#/components/schemas/UpsertTemplateMutation'
        - $ref: '#/components/schemas/DeleteMutation'
      discriminator:
        propertyName: kind
        mapping:
          upsertRecord: '#/components/schemas/UpsertRecordMutation'
          upsertEvent: '#/components/schemas/UpsertEventMutation'
          upsertTag: '#/components/schemas/UpsertTagMutation'
          upsertTemplate: '#/components/schemas/UpsertTemplateMutation'
          delete: '#/components/schemas/DeleteMutation'
    SyncBatchRequest:
      type: object
      required: [deviceId, mutations]
      properties:
        deviceId:
          type: string
          format: uuid
          description: Must match an API key's bound device when API-key authenticated.
        atomic:
          type: boolean
          default: false
        mutations:
          type: array
          minItems: 1
          maxItems: 20
          items:
            $ref: '#/components/schemas/SyncMutation'
      additionalProperties: false
    MutationStatus:
      type: string
      enum: [succeeded, failed]
    SyncMutationResult:
      type: object
      required: [clientMutationId, status]
      properties:
        clientMutationId:
          $ref: '#/components/schemas/MutationId'
        status:
          $ref: '#/components/schemas/MutationStatus'
        resourceType:
          $ref: '#/components/schemas/SyncResourceType'
        resourceId:
          $ref: '#/components/schemas/ResourceIdentifier'
        revision:
          type: integer
          format: int64
          minimum: 1
        etag:
          type: string
        problem:
          $ref: '#/components/schemas/Problem'
      additionalProperties: false
    SyncBatchResponse:
      type: object
      required: [results]
      properties:
        results:
          type: array
          minItems: 1
          maxItems: 20
          items:
            $ref: '#/components/schemas/SyncMutationResult'
      additionalProperties: false